+++ pkg/mesh-networking/apply/applier.go
@@ -66,7 +66,7 @@ func (v *applier) Apply(ctx context.Context, input input.Snapshot) {
 	for _, trafficPolicy := range input.TrafficPolicies().List() {
 		trafficPolicy.Status = networkingv1alpha2.TrafficPolicyStatus{
 			State:              networkingv1alpha2.ApprovalState_ACCEPTED,
-			ObservedGeneration: trafficPolicy.Generation,
+			ObservedGeneration: uint32(trafficPolicy.Generation),
 			TrafficTargets:     map[string]*networkingv1alpha2.ApprovalStatus{},
 		}
 	}
@@ -76,7 +76,7 @@ func (v *applier) Apply(ctx context.Context, input input.Snapshot) {
 	for _, accessPolicy := range input.AccessPolicies().List() {
 		accessPolicy.Status = networkingv1alpha2.AccessPolicyStatus{
 			State:              networkingv1alpha2.ApprovalState_ACCEPTED,
-			ObservedGeneration: accessPolicy.Generation,
+			ObservedGeneration: uint32(accessPolicy.Generation),
 			TrafficTargets:     map[string]*networkingv1alpha2.ApprovalStatus{},
 		}
 	}
@@ -86,7 +86,7 @@ func (v *applier) Apply(ctx context.Context, input input.Snapshot) {
 	for _, failoverService := range input.FailoverServices().List() {
 		failoverService.Status = networkingv1alpha2.FailoverServiceStatus{
 			State:              networkingv1alpha2.ApprovalState_ACCEPTED,
-			ObservedGeneration: failoverService.Generation,
+			ObservedGeneration: uint32(failoverService.Generation),
 			Meshes:             map[string]*networkingv1alpha2.ApprovalStatus{},
 		}
 	}
@@ -97,19 +97,19 @@ func (v *applier) Apply(ctx context.Context, input input.Snapshot) {
 		if virtualMesh.Status.State != networkingv1alpha2.ApprovalState_ACCEPTED {
 			continue
 		}
-		virtualMesh.Status.ObservedGeneration = virtualMesh.Generation
+		virtualMesh.Status.ObservedGeneration = uint32(virtualMesh.Generation)
 		virtualMesh.Status.Meshes = map[string]*networkingv1alpha2.ApprovalStatus{}
 	}
 
 	// update traffictarget, trafficpolicy, and accesspolicy statuses
 	for _, trafficTarget := range input.TrafficTargets().List() {
-		trafficTarget.Status.ObservedGeneration = trafficTarget.Generation
+		trafficTarget.Status.ObservedGeneration = uint32(trafficTarget.Generation)
 		trafficTarget.Status.AppliedTrafficPolicies = validateAndReturnApprovedTrafficPolicies(ctx, input, reporter, trafficTarget)
 		trafficTarget.Status.AppliedAccessPolicies = validateAndReturnApprovedAccessPolicies(ctx, input, reporter, trafficTarget)
 	}
 
 	for _, mesh := range input.Meshes().List() {
-		mesh.Status.ObservedGeneration = mesh.Generation
+		mesh.Status.ObservedGeneration = uint32(mesh.Generation)
 		mesh.Status.AppliedFailoverServices = validateAndReturnApprovedFailoverServices(ctx, input, reporter, mesh)
 		mesh.Status.AppliedVirtualMesh = validateAndReturnVirtualMesh(ctx, input, reporter, mesh)
 	}
@@ -474,7 +474,7 @@ func getAppliedTrafficPolicies(
 		appliedPolicies = append(appliedPolicies, &discoveryv1alpha2.TrafficTargetStatus_AppliedTrafficPolicy{
 			Ref:                ezkube.MakeObjectRef(policy),
 			Spec:               &policy.Spec,
-			ObservedGeneration: policy.Generation,
+			ObservedGeneration: uint32(policy.Generation),
 		})
 	}
 	return appliedPolicies
@@ -488,7 +488,7 @@ func getAppliedTrafficPolicies(
 // Finally, policies which are rejected and modified
 func sortTrafficPoliciesByAcceptedDate(trafficTarget *discoveryv1alpha2.TrafficTarget, trafficPolicies networkingv1alpha2.TrafficPolicySlice) {
 	isUpToDate := func(tp *networkingv1alpha2.TrafficPolicy) bool {
-		return tp.Status.ObservedGeneration == tp.Generation
+		return int64(tp.Status.ObservedGeneration) == tp.Generation
 	}
 
 	sort.SliceStable(trafficPolicies, func(i, j int) bool {
@@ -544,7 +544,7 @@ func getAppliedAccessPolicies(
 		appliedPolicies = append(appliedPolicies, &discoveryv1alpha2.TrafficTargetStatus_AppliedAccessPolicy{
 			Ref:                ezkube.MakeObjectRef(policy),
 			Spec:               &policy.Spec,
-			ObservedGeneration: policy.Generation,
+			ObservedGeneration: uint32(policy.Generation),
 		})
 	}
 
@@ -565,7 +565,7 @@ func getAppliedVirtualMesh(
 				return &discoveryv1alpha2.MeshStatus_AppliedVirtualMesh{
 					Ref:                ezkube.MakeObjectRef(vMesh),
 					Spec:               &vMesh.Spec,
-					ObservedGeneration: vMesh.Generation,
+					ObservedGeneration: uint32(vMesh.Generation),
 				}
 			}
 		}
@@ -588,7 +588,7 @@ func getAppliedFailoverServices(
 			appliedFailoverServices = append(appliedFailoverServices, &discoveryv1alpha2.MeshStatus_AppliedFailoverService{
 				Ref:                ezkube.MakeObjectRef(failoverService),
 				Spec:               &failoverService.Spec,
-				ObservedGeneration: failoverService.Generation,
+				ObservedGeneration: uint32(failoverService.Generation),
 			})
 		}
 	}
